requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
emoji==1.7.0
filelock==3.7.1
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
fonttools==4.34.2
greenlet==1.1.2
gunicorn==20.1.0
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
llvmlite==0.38.1
MarkupSafe==2.1.1
matplotlib==3.5.2
numba==0.55.2
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
SQLAlchemy==1.4.39
threadpoolctl==3.1.0
tokenizers==0.12.1
torch
torchvision
tqdm==4.64.0
transformers==4.20.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
